+++ src/interfaces/wallet.cpp
@@ -393,6 +393,7 @@ namespace {
         OutputType getDefaultAddressType() override {
             return m_wallet.m_default_address_type;
         }
+        bool canGetAddresses() override { return m_wallet.CanGetAddresses(); }
         bool IsWalletFlagSet(uint64_t flag) override {
             return m_wallet.IsWalletFlagSet(flag);
         }