+++ src/interfaces/wallet.h
@@ -222,6 +222,9 @@ public:
     // Return whether HD enabled.
     virtual bool hdEnabled() = 0;
 
+    // Return whether the wallet is blank.
+    virtual bool canGetAddresses() = 0;
+
     // Check if a certain wallet flag is set.
     virtual bool IsWalletFlagSet(uint64_t flag) = 0;