+++ src/rpc/client.cpp
@@ -157,6 +157,7 @@ static const CRPCConvertParam vRPCConvertParams[] = {
     {"rescanblockchain", 0, "start_height"},
     {"rescanblockchain", 1, "stop_height"},
     {"createwallet", 1, "disable_private_keys"},
+    {"createwallet", 2, "blank"},
 };
 
 class CRPCConvertTable {